Move definitions, simpler names for languages
