+++ client/src/definitions/graph.ts
