+++ client/src/definitions/values.ts
