--- client/src/editor.css
@@ -1,34 +1,0 @@
-div.output.rendered {
-    display: block;
-}
-
-div.output {
-    display: none;
-}
-
-div.editor.editable {
-    display: block;
-}
-
-div.editor {
-    display: none;
-}
-
-div.cell{
-    border-left-style: solid;
-    border-left-width: 7px;
-    padding: 5px 5px 5px 15px;
-    width:100%;
-}
-
-div.cell:nth-child(odd){
-   border-left-color: gainsboro; 
-}
-
-div.cell:nth-child(even){
-    border-left-color: grey;
-}
-
-div.controls {
-	float:right;
-}
