--- client/src/languagePlugins/markdown/editor.css
@@ -1,24 +1,0 @@
-div.editor {
-    border-left-style: double;
-    border-left-color: grey;
-    padding: 5px;
-    margin: 10px;
-    height: 50vh;
-    /* overflow: scroll; */
-}
-
-div.output.rendered {
-    display: block;
-}
-
-div.output {
-    display: none;
-}
-
-div.editor.editable {
-    display: block;
-}
-
-div.editor {
-    display: none;
-}
